lambda/shared/requirements.txt
cryptography==41.0.7
PyJWT==2.8.0